+++ Tests/MapboxMapsTests/Foundation/Style/MapStyleReconcilerTests.swift
@@ -56,11 +56,7 @@ final class MapStyleReconcilerTests: XCTestCase {
         let json = """
         {"foo": "bar"}
         """
-        me.mapStyle = .init(json: json, importConfigurations: [
-            .init(importId: "foo", config: [
-                "bar": "baz"
-            ])
-        ])
+        me.mapStyle = .init(json: json, configuration: JSONObject(rawValue: ["bar": "baz"])!)
 
         XCTAssertEqual(styleManager.setStyleJSONStub.invocations.count, 1)
         let params = try XCTUnwrap(styleManager.setStyleJSONStub.invocations.last).parameters
@@ -73,7 +69,7 @@ final class MapStyleReconcilerTests: XCTestCase {
 
         let inv = styleManager.setStyleImportConfigPropertyForImportIdStub.invocations
         XCTAssertEqual(inv.count, 1)
-        XCTAssertEqual(inv.last?.parameters.importId, "foo")
+        XCTAssertEqual(inv.last?.parameters.importId, "basemap")
         XCTAssertEqual(inv.last?.parameters.config, "bar")
         XCTAssertEqual(inv.last?.parameters.value as? String, "baz")
     }
@@ -83,11 +79,7 @@ final class MapStyleReconcilerTests: XCTestCase {
             self.styleManager.isStyleLoadedStub.defaultReturnValue = false
         }
 
-        me.mapStyle = .init(uri: .streets, importConfigurations: [
-            .init(importId: "foo", config: [
-                "bar": "baz"
-            ])
-        ])
+        me.mapStyle = .init(uri: .streets, configuration: JSONObject(rawValue: ["bar": "baz"])!)
 
         XCTAssertEqual(styleManager.setStyleURIStub.invocations.count, 1)
         let params = try XCTUnwrap(styleManager.setStyleURIStub.invocations.last).parameters
@@ -100,7 +92,7 @@ final class MapStyleReconcilerTests: XCTestCase {
 
         let inv = styleManager.setStyleImportConfigPropertyForImportIdStub.invocations
         XCTAssertEqual(inv.count, 1)
-        XCTAssertEqual(inv.last?.parameters.importId, "foo")
+        XCTAssertEqual(inv.last?.parameters.importId, "basemap")
         XCTAssertEqual(inv.last?.parameters.config, "bar")
         XCTAssertEqual(inv.last?.parameters.value as? String, "baz")
     }
@@ -115,12 +107,8 @@ final class MapStyleReconcilerTests: XCTestCase {
             callbacks = invoc.parameters.callbacks
         }
 
-        me.mapStyle = MapStyle(uri: .outdoors, importConfigurations: [
-            .init(importId: "foo-1", config: ["k-1": "v-1", "a": "b"])
-        ])
-        me.mapStyle = MapStyle(uri: .streets, importConfigurations: [
-            .init(importId: "foo-2", config: ["k-2": "v-2"])
-        ])
+        me.mapStyle = .init(uri: .outdoors, configuration: JSONObject(rawValue: ["k-1": "v-1", "a": "b"])!)
+        me.mapStyle = .init(uri: .streets, configuration: JSONObject(rawValue: ["k-2": "v-2"])!)
 
         XCTAssertEqual(styleManager.setStyleURIStub.invocations.map(\.parameters.value), [
             StyleURI.outdoors.rawValue,
@@ -133,7 +121,7 @@ final class MapStyleReconcilerTests: XCTestCase {
         // the first style update is skipped.
         let inv = styleManager.setStyleImportConfigPropertyForImportIdStub.invocations
         XCTAssertEqual(inv.count, 1)
-        XCTAssertEqual(inv.last?.parameters.importId, "foo-2")
+        XCTAssertEqual(inv.last?.parameters.importId, "basemap")
         XCTAssertEqual(inv.last?.parameters.config, "k-2")
         XCTAssertEqual(inv.last?.parameters.value as? String, "v-2")
     }
@@ -204,13 +192,9 @@ final class MapStyleReconcilerTests: XCTestCase {
             self.simulateLoad(callbacks: invoc.parameters.callbacks, result: .success)
             self.styleManager.setStyleImportConfigPropertyForImportIdStub.reset()
         }
-        me.mapStyle = MapStyle(uri: .outdoors, importConfigurations: [
-            .init(importId: "foo-1", config: ["k-1": "v-1", "a": "b"])
-        ])
+        me.mapStyle = MapStyle(uri: .outdoors, configuration: JSONObject(rawValue: ["k-1": "v-1", "a": "b"])!)
 
-        let s2 = MapStyle(uri: .outdoors, importConfigurations: [
-            .init(importId: "foo-1", config: ["k-2": "v-2"])
-        ])
+        let s2 = MapStyle(uri: .outdoors, configuration: JSONObject(rawValue: ["k-2": "v-2"])!)
 
         var count = 0
         me.loadStyle(s2, transition: nil) { error in
@@ -221,53 +205,41 @@ final class MapStyleReconcilerTests: XCTestCase {
 
         let inv = styleManager.setStyleImportConfigPropertyForImportIdStub.invocations
         XCTAssertEqual(inv.count, 1)
-        XCTAssertEqual(inv.last?.parameters.importId, "foo-1")
+        XCTAssertEqual(inv.last?.parameters.importId, "basemap")
         XCTAssertEqual(inv.last?.parameters.config, "k-2")
         XCTAssertEqual(inv.last?.parameters.value as? String, "v-2")
     }
 
     func testStyleImportsReconcileFromNil() {
-        MapStyleReconciler.reconcileStyleImports(
+            MapStyleReconciler.reconcileBasemapConfiguration(
             from: nil,
-            to: [
-                StyleImportConfiguration(
-                    importId: "foo",
-                    config: ["bar": "baz"])
-            ],
+            to: .init(rawValue: ["bar": "baz"])!,
             styleManager: styleManager)
 
         let inv = styleManager.setStyleImportConfigPropertyForImportIdStub.invocations
         XCTAssertEqual(inv.count, 1)
-        XCTAssertEqual(inv.last?.parameters.importId, "foo")
+        XCTAssertEqual(inv.last?.parameters.importId, "basemap")
         XCTAssertEqual(inv.last?.parameters.config, "bar")
         XCTAssertEqual(inv.last?.parameters.value as? String, "baz")
     }
 
     func testStyleImportsReconcilePartialUpdate() {
-        MapStyleReconciler.reconcileStyleImports(
-            from: [
-                StyleImportConfiguration(
-                    importId: "foo",
-                    config: ["bar": "baz"]),
-                StyleImportConfiguration(
-                    importId: "x",
-                    config: ["y": "z"])
-            ],
-            to: [
-                StyleImportConfiguration(
-                    importId: "foo",
-                    config: [
-                        "bar": "baz",
-                        "qux": "quux"
-                    ])
-            ],
+        MapStyleReconciler.reconcileBasemapConfiguration(
+            from: JSONObject(rawValue: [
+                "bar": "baz",
+                "x": "y"
+            ])!,
+            to: JSONObject(rawValue: [
+                "bar": "foo",
+                "x": "y"
+            ])!,
             styleManager: styleManager)
 
         let inv = styleManager.setStyleImportConfigPropertyForImportIdStub.invocations
         XCTAssertEqual(inv.count, 1)
-        XCTAssertEqual(inv.last?.parameters.importId, "foo")
-        XCTAssertEqual(inv.last?.parameters.config, "qux")
-        XCTAssertEqual(inv.last?.parameters.value as? String, "quux")
+        XCTAssertEqual(inv.last?.parameters.importId, "basemap")
+        XCTAssertEqual(inv.last?.parameters.config, "bar")
+        XCTAssertEqual(inv.last?.parameters.value as? String, "foo")
     }
 
     func testIsStyleRootLoaded() {