+++ Tests/MapboxMapsTests/SwiftUI/MapBasicCoordinatorTests.swift
@@ -32,16 +32,16 @@ final class MapBasicCoordinatorTests: XCTestCase {
     func testStyleURI() {
         update(with: MapDependencies(mapStyle: .light))
 
-        var loadMethod = mapView.style.mapStyle?.loadMethod
-        if case let .uri(styleURI) = loadMethod {
+        var styleData = mapView.style.mapStyle?.data
+        if case let .uri(styleURI) = styleData {
             XCTAssertEqual(styleURI.rawValue, "mapbox://styles/mapbox/light-v11")
         } else {
             XCTFail("Failed to update mapStyle")
         }
 
         update(with: MapDependencies(mapStyle: .dark))
-        loadMethod = mapView.style.mapStyle?.loadMethod
-        if case let .uri(styleURI) = loadMethod {
+        styleData = mapView.style.mapStyle?.data
+        if case let .uri(styleURI) = styleData {
             XCTAssertEqual(styleURI.rawValue, "mapbox://styles/mapbox/dark-v11")
         } else {
             XCTFail("Failed to update mapStyle")
@@ -260,7 +260,13 @@ extension PerformanceStatistics {
         PerformanceStatistics(
             collectionDurationMillis: 1000,
             mapRenderDurationStatistics: DurationStatistics(maxMillis: 0, medianMillis: 0),
-            cumulativeStatistics: CumulativeRenderingStatistics(drawCalls: nil, textureBytes: nil, vertexBytes: nil),
+            cumulativeStatistics: CumulativeRenderingStatistics(
+                drawCalls: nil,
+                textureBytes: nil,
+                vertexBytes: nil,
+                graphicsPrograms: nil,
+                graphicsProgramsCreationTimeMillis: nil
+            ),
             perFrameStatistics: PerFrameRenderingStatistics(
                 topRenderGroups: [],
                 topRenderLayers: [],